ai-job-finder/requirements.txt
Flask
google-genai
python-dotenv
fastapi